type fixes
